+++ src/state/state.hpp
@@ -54,7 +54,7 @@ public:
   State(Board board): board(board) {};
   State(Board board, int player): board(board), player(player) {};
 
-  int evaluate(bool self);
+  double evaluate(bool self);
   State* next_state(Move move);
   void get_legal_actions();
   std::string encode_output();
